image-metadata-app/requirements.txt
openai>=1.40.0
Pillow>=10.0.0
jsonschema>=4.17.0
python-dotenv>=1.0.0
